Update variables to be float

+++ Particle.h
@@ -1,9 +1,9 @@
 #pragma once
 struct Particle {
-    double x;
-    double y;
-    double m;
-    double r;
-    double vx;
-    double vy;
+    float x;
+    float y;
+    float m;
+    float r;
+    float vx;
+    float vy;
 };
